Pull request: to Have Decimal Values on Altitude recorded on tracks

+++ app/src/main/java/ch/bailu/aat/gpx/GpxDataWrapper.java
@@ -65,7 +65,7 @@ public class GpxDataWrapper implements GpxDeltaPointInterface, GpxBigDeltaInterf
     }
 
     @Override
-    public short getAltitude() {
+    public double getAltitude() {
         return point.getAltitude();
     }
