+++ app/src/main/java/ch/bailu/aat/services/location/NewAltitudeFromBarometer.java
@@ -19,7 +19,7 @@ public class NewAltitudeFromBarometer extends LocationStackChainedItem {
                 scontext.getSensorService().getInformationOrNull(InfoID.BAROMETER_SENSOR);
 
         if (info != null) {
-            short altitude = info.getAltitude();
+            double altitude = info.getAltitude();
             l.setAltitude(altitude);
         }
